+++ wxc_sdk/telephony/calls.py
@@ -560,6 +560,7 @@ class CallsApi(ApiChild, base='telephony/calls'):
     def call_history(self, history_type: Union[str, HistoryType] = None) -> Generator[CallHistoryRecord, None, None]:
         """
         List Call History
+
         Get the list of call history records for the user. A maximum of 20 call history records per type (placed,
         missed, received) are returned.
 
@@ -571,3 +572,33 @@ class CallsApi(ApiChild, base='telephony/calls'):
         params = history_type and {'type': enum_str(history_type)} or None
         url = self.ep('history')
         return self.session.follow_pagination(url=url, model=CallHistoryRecord, params=params)
+
+    def mute(self, call_id: str):
+        """
+        Mute
+
+        Mute a call.
+
+        :param call_id: The call identifier of the call to mute.
+        :type call_id: str
+        :rtype: None
+        """
+        body = dict()
+        body['callId'] = call_id
+        url = self.ep('mute')
+        super().post(url, json=body)
+
+    def unmute(self, call_id: str):
+        """
+        Unmute
+
+        Unmute a call.
+
+        :param call_id: The call identifier of the call to unmute.
+        :type call_id: str
+        :rtype: None
+        """
+        body = dict()
+        body['callId'] = call_id
+        url = self.ep('unmute')
+        super().post(url, json=body)